feat(provider): make openai, anthropic, google, and amazon-bedrock providers optional via build tags
Every model provider was wired in through a single chokepoint
(pkg/model/provider/factory.go), so the heavy provider SDKs only entered the
dependency graph through that one file. This makes it cheap to compile a
provider out for projects embedding docker-agent that don't need it.

Split the monolithic factory into a shared core plus one self-registering file
per optional provider, each gated by a negative build tag:

  - docker_agent_no_openai    drops github.com/openai/openai-go
  - docker_agent_no_anthropic drops github.com/anthropics/anthropic-sdk-go
  - docker_agent_no_google    drops google.golang.org/genai and the Vertex AI
                              / cloud auth stack
  - docker_agent_no_bedrock   drops the github.com/aws/aws-sdk-go-v2 stack
                              (the largest provider-specific dependency tree)

All providers are included by default; the dmr provider and rule-based router
are always compiled in. Requesting a model whose provider was compiled out
fails at construction time with a clear 'not compiled into this build' error.

Build tags are global to a build (not scoped per module), so the tags are
prefixed with docker_agent_ to avoid colliding with an embedding project's own
tags. The rationale and the full tag list are documented in the package doc so
downstream consumers can discover them via go doc / pkg.go.dev.

Tests that build a real OpenAI client to exercise generic clone/custom-provider
logic are tagged !docker_agent_no_openai; the provider-agnostic logic tests
stay in the default build.
