+++ pkg/model/provider/clone_test.go
@@ -1,3 +1,5 @@
+//go:build !js && !docker_agent_no_openai
+
 package provider
 
 import (
