+++ pkg/model/provider/factory.go
@@ -9,14 +9,9 @@ import (
 
 	"github.com/docker/docker-agent/pkg/config/latest"
 	"github.com/docker/docker-agent/pkg/environment"
-	"github.com/docker/docker-agent/pkg/model/provider/anthropic"
-	"github.com/docker/docker-agent/pkg/model/provider/bedrock"
 	"github.com/docker/docker-agent/pkg/model/provider/dmr"
-	"github.com/docker/docker-agent/pkg/model/provider/gemini"
-	"github.com/docker/docker-agent/pkg/model/provider/openai"
 	"github.com/docker/docker-agent/pkg/model/provider/options"
 	"github.com/docker/docker-agent/pkg/model/provider/rulebased"
-	"github.com/docker/docker-agent/pkg/model/provider/vertexai"
 )
 
 // createRuleBasedRouter creates a rule-based routing provider.
@@ -71,9 +66,18 @@ func createDirectProvider(ctx context.Context, cfg *latest.ModelConfig, env envi
 	factory, ok := providerFactories[providerType]
 	if !ok {
 		slog.ErrorContext(ctx, "Unknown provider type", "type", providerType)
-		return nil, fmt.Errorf("unknown provider type: %s", providerType)
+		return nil, unknownProviderError(providerType)
 	}
 	return factory(ctx, enhancedCfg, env, opts...)
+}
+
+// unknownProviderError builds the error returned when no factory is registered
+// for providerType. The provider may be genuinely unknown, or it may have been
+// compiled out via a `no_<provider>` build tag (openai/anthropic/google/bedrock
+// are optional). The message hints at the latter so trimmed builds are easy to
+// diagnose.
+func unknownProviderError(providerType string) error {
+	return fmt.Errorf("unknown provider type %q (it may not be compiled into this build)", providerType)
 }
 
 // providerFactory builds a Provider from a fully-resolved ModelConfig.
@@ -82,53 +86,31 @@ func createDirectProvider(ctx context.Context, cfg *latest.ModelConfig, env envi
 type providerFactory func(ctx context.Context, cfg *latest.ModelConfig, env environment.Provider, opts ...options.Opt) (Provider, error)
 
 // providerFactories maps a resolved provider type (the value returned by
-// resolveProviderType) to its constructor. The map is package-private but
-// modifiable; tests must restore the original entries with t.Cleanup.
-var providerFactories = map[string]providerFactory{
-	"openai":                 openaiFactory,
-	"openai_chatcompletions": openaiFactory,
-	"openai_responses":       openaiFactory,
-	"anthropic":              anthropicFactory,
-	"google":                 googleFactory,
-	"dmr":                    dmrFactory,
-	"amazon-bedrock":         bedrockFactory,
-}
+// resolveProviderType) to its constructor.
+//
+// Entries are contributed by per-provider files via registerProviderFactory in
+// their init(). The optional providers (openai, anthropic, google,
+// amazon-bedrock) each live behind a `no_<provider>` build tag so embedders can
+// drop them — and their transitive SDK dependencies — from a build. The map is
+// package-private but modifiable; tests must restore the original entries with
+// t.Cleanup.
+var providerFactories = map[string]providerFactory{}
 
-func openaiFactory(ctx context.Context, cfg *latest.ModelConfig, env environment.Provider, opts ...options.Opt) (Provider, error) {
-	return openai.NewClient(ctx, cfg, env, opts...)
-}
-
-func anthropicFactory(ctx context.Context, cfg *latest.ModelConfig, env environment.Provider, opts ...options.Opt) (Provider, error) {
-	return anthropic.NewClient(ctx, cfg, env, opts...)
-}
-
-func googleFactory(ctx context.Context, cfg *latest.ModelConfig, env environment.Provider, opts ...options.Opt) (Provider, error) {
-	// Route non-Gemini models on Vertex AI (Model Garden) through the
-	// vertexai package, which picks the right endpoint per publisher.
-	if vertexai.IsModelGardenConfig(cfg) {
-		return vertexClientFactory(ctx, cfg, env, opts...)
+// registerProviderFactory records f under name in providerFactories. It is
+// called from the init() of each per-provider file; duplicate registration of
+// the same name panics to surface wiring mistakes at startup.
+func registerProviderFactory(name string, f providerFactory) {
+	if _, exists := providerFactories[name]; exists {
+		panic("provider: duplicate factory registration for " + name)
 	}
-	return geminiClientFactory(ctx, cfg, env, opts...)
+	providerFactories[name] = f
 }
 
-// geminiClientFactory and vertexClientFactory are the inner constructors used
-// by googleFactory. They are package-level variables (rather than direct
-// references to gemini.NewClient / vertexai.NewClient) so that tests can swap
-// them with fakes via t.Cleanup and assert that googleFactory routes correctly
-// based on vertexai.IsModelGardenConfig — without spinning up real clients.
-var (
-	geminiClientFactory providerFactory = func(ctx context.Context, cfg *latest.ModelConfig, env environment.Provider, opts ...options.Opt) (Provider, error) {
-		return gemini.NewClient(ctx, cfg, env, opts...)
-	}
-	vertexClientFactory providerFactory = func(ctx context.Context, cfg *latest.ModelConfig, env environment.Provider, opts ...options.Opt) (Provider, error) {
-		return vertexai.NewClient(ctx, cfg, env, opts...)
-	}
-)
+//nolint:gochecknoinits // Intentional: providers self-register so optional ones can be dropped via build tags.
+func init() {
+	registerProviderFactory("dmr", dmrFactory)
+}
 
 func dmrFactory(ctx context.Context, cfg *latest.ModelConfig, _ environment.Provider, opts ...options.Opt) (Provider, error) {
 	return dmr.NewClient(ctx, cfg, opts...)
-}
-
-func bedrockFactory(ctx context.Context, cfg *latest.ModelConfig, env environment.Provider, opts ...options.Opt) (Provider, error) {
-	return bedrock.NewClient(ctx, cfg, env, opts...)
 }
