+++ pkg/model/provider/provider.go
@@ -1,6 +1,6 @@
 // Package provider builds and dispatches to LLM provider clients.
 //
-// The package is organised across four files:
+// The package is organised across several files:
 //
 //   - provider.go (this file): the public Provider interfaces and the entry
 //     points [New] and [NewWithModels] that callers use to construct a
@@ -11,9 +11,50 @@
 //   - defaults.go: pure config-merging logic that fills in defaults from
 //     custom providers, built-in aliases, and model-specific rules
 //     (thinking budget, interleaved thinking, ...).
-//   - factory.go: dispatch from a resolved provider type to the concrete
-//     client constructor (openai, anthropic, google, dmr, amazon-bedrock,
-//     vertex AI), plus the rule-based router.
+//   - factory.go: shared dispatch from a resolved provider type to the
+//     concrete client constructor, plus the always-available dmr provider
+//     and the rule-based router.
+//   - factory_<name>.go: one file per optional provider (openai, anthropic,
+//     google, amazon-bedrock); each registers itself with the dispatch table
+//     and is gated by a build tag (see "Build tags" below).
+//
+// # Build tags
+//
+// The openai, anthropic, google and amazon-bedrock providers are optional.
+// Each lives behind a negative build tag so a project embedding docker-agent
+// can compile a provider out — together with its transitive SDK dependencies —
+// to shrink the binary and dependency graph. All providers are included by
+// default; pass the relevant tag(s) to opt out.
+//
+// Build tags are global to a build, not scoped per module: a tag set by the
+// top-level project applies to every dependency too. The tags are therefore
+// prefixed with "docker_agent_" so an embedding project can use its own build
+// tags (even a plain "no_openai") without accidentally toggling these
+// providers.
+//
+// The available tags are:
+//
+//   - docker_agent_no_openai: drop the OpenAI provider (github.com/openai/openai-go).
+//   - docker_agent_no_anthropic: drop the Anthropic provider
+//     (github.com/anthropics/anthropic-sdk-go). The google provider's Vertex
+//     Model Garden support also imports the anthropic package, so the
+//     dependency is only fully removed when combined with docker_agent_no_google.
+//   - docker_agent_no_google: drop the Google provider (google.golang.org/genai, the
+//     Vertex AI / cloud auth stack, and — via Vertex Model Garden — the
+//     anthropic and openai SDKs). Vertex AI is unsupported either way under
+//     js/wasm.
+//   - docker_agent_no_bedrock: drop the Amazon Bedrock provider
+//     (the github.com/aws/aws-sdk-go-v2 stack), the largest provider-specific
+//     dependency tree.
+//
+// For example, to build without Bedrock and OpenAI:
+//
+//	go build -tags 'docker_agent_no_bedrock docker_agent_no_openai' ...
+//
+// Requesting a model whose provider was compiled out fails at construction
+// time with a clear "not compiled into this build" error rather than at
+// compile time. The dmr provider and the rule-based router are always
+// compiled in (except under js/wasm, which has its own slim factory).
 package provider
 
 import (
